@@ -1,9 +1,9 @@
 package nz.co.vilemob.cointosser
 
-import android.arch.lifecycle.Observer
-import android.arch.lifecycle.ViewModelProviders
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.Observer
+import androidx.lifecycle.ViewModelProviders
 import kotlinx.android.synthetic.main.activity_coin_tosser.*
 
 class CoinTosserActivity : AppCompatActivity() {
@@ -15,7 +15,7 @@ class CoinTosserActivity : AppCompatActivity() {
         val viewModel = ViewModelProviders.of(this).get(CoinTosserViewModel::class.java)
         viewModel.liveState.observe(this, Observer {
             it?.apply {
-                statusTextView.text = when{
+                statusTextView.text = when {
                     isTossing -> getString(R.string.tossing)
                     isHeads -> getString(R.string.heads)
                     else -> getString(R.string.tails)
